cleanup constructor for the metal device
